requirements.txt
GDAL==3.6.4
numpy==1.25.0
protobuf==4.23.3
pypdf2==3.0.*
langchain===0.0.216
unstructured
chromadb
tiktoken
python-dotenv
openai